Added middleware for Subtasks

+++ routes/web.php
@@ -23,7 +23,7 @@ Auth::routes();
 
     Route::post('/task','TaskController@store')->name('Task.store')->middleware('isRelatedtoTask');
     Route::get('/Task/{id}', 'TaskController@show')->name('Task.show')->middleware('isRelatedtoTask');
-    Route::get('/task/create','TaskController@create')->name('Task.create')->middleware('isRelatedtoTask');
+    Route::get('/task/create','TaskController@create')->name('Task.create');
     Route::get('/task/{id}/edit','TaskController@edit')->name('Task.edit')->middleware('isRelatedtoTask')->middleware('isAdmin');
     Route::patch('/task/update', 'TaskController@update')->name('Task.update')->middleware('isRelatedtoTask')->middleware('isAdmin');
     Route::delete('/task/{id}/delete', 'TaskController@destroy')->name('Task.destroy')->middleware('isRelatedtoTask')->middleware('isAdmin');
@@ -36,8 +36,8 @@ Auth::routes();
     Route::get('/task/{id}/subtasks','TaskController@showsubtasks')->name('Task.subtasks')->middleware('isRelatedtoTask');
 
     Route::post('/subtask','SubtaskController@store')->name('Subtask.store');
-    Route::get('/subtask/{id}', 'SubtaskController@show')->name('Subtask.show');
-    Route::get('/subtask/create/{id}','SubtaskController@create')->name('Subtask.create');
-    Route::get('/subtask/{id}/edit','SubtaskController@edit')->name('Subtask.edit');
+    Route::get('/subtask/{id}', 'SubtaskController@show')->name('Subtask.show')->middleware('isRelatedtoSubtask');
+    Route::get('/subtask/create/{id}','SubtaskController@create')->name('Subtask.create')->middleware('isAdmin');
+    Route::get('/subtask/{id}/edit','SubtaskController@edit')->name('Subtask.edit')->middleware('isAdminSubtask');
     Route::patch('/subtask/update', 'SubtaskController@update')->name('Subtask.update');
-    Route::delete('/subtask/{id}/delete', 'SubtaskController@destroy')->name('Subtask.destroy');
+    Route::delete('/subtask/{id}/delete', 'SubtaskController@destroy')->name('Subtask.destroy')->middleware('isAdminSubtask');
